No deep-linking between "primary" modules (app, sim, micro); this is borked with circular dependencies;
