+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "bounce-back-railway",
-  "version": "1.0.1",
+  "version": "1.0.2",
   "description": "A JavaScript visual prototype implementation for a mini-train microcontroller that auto-reverses at the ends of a straight track, with any number of intermediate stops. Perfect for model railroaders looking for a simple and fun automation solution.",
   "main": "./dist/trolley-controller.bundle.js",
   "exports": {
