--- src/app/constants/index.ts
@@ -1,4 +1,0 @@
-export * from './DirectionTypes';
-export * from './PinAssignments';
-export * from './SensorTypes';
-export * from './Constants';
