+++ src/app/libs/mgrs/ControlManager.ts
@@ -1,12 +1,14 @@
-import { uint10_t, uint8_t } from '../../interfaces';
 import {
-  DASHBORD_CHBX,
-  SIGNAL_CHBX,
-  STATION_CHBX,
-  HALT_BTN, PAUSE_BTN, POWER_BTN, REVERSE_BTN, SENSOR_26TH_AVE_XING, SENSOR_ATWATER_XING, SENSOR_LT_LA, SENSOR_MIDDLE_BURBANK, SENSOR_RT_CLAREMONT, SENSOR_SOUTH_GATE_XING, SENSOR_VOLTS_ALL_CLEAR, SPEED_CONTROL,
+  uint10_t,
+  uint8_t,
   uint10_MAX,
-  CONTROL_PANEL_CHBX,
-} from '../../constants';
+
+  CONTROL_PANEL_CHBX, DASHBORD_CHBX, SIGNAL_CHBX, STATION_CHBX,
+  HALT_BTN, PAUSE_BTN, POWER_BTN, REVERSE_BTN,
+  SPEED_CONTROL,
+  SENSOR_26TH_AVE_XING, SENSOR_ATWATER_XING, SENSOR_LT_LA, SENSOR_MIDDLE_BURBANK, SENSOR_RT_CLAREMONT, SENSOR_SOUTH_GATE_XING,
+  SENSOR_VOLTS_ALL_CLEAR,
+} from '../../../common';
 import { Serial } from '../../../microcontroller';
 
 interface CurrentState {
