+++ src/common/configs/StationsUtils.ts
@@ -1,6 +1,8 @@
-import { SensorTypes } from '../app/constants';
+import { SensorTypes } from '../constants/SensorTypes';
 import { stations } from './Stations';
 
 const getByType = (filterType: SensorTypes) => stations.filter(({ type }: { type: SensorTypes}) => type === filterType);
+
 export const getSignals = getByType.bind(null, SensorTypes.SIGNAL);
+
 export const getStations = getByType.bind(null, SensorTypes.STATION);
