+++ src/common/constants/MicrocontrollerConstants.ts
@@ -1,4 +1,4 @@
-import { milliseconds } from '../../app/interfaces';
+import { milliseconds } from '../interfaces/CoreTypes';
 
 /** (ms/tick) i.e. tick duration */
 export const CLOCK_SPEED: milliseconds = 5;
