+++ src/common/constants/SensorTypes.ts
