+++ src/common/interfaces/CoreTypes.ts
