--- src/microcontroller/constants/index.ts
@@ -1,2 +1,0 @@
-export * from './ArduinoUno';
-export * from './Constants';
