+++ src/microcontroller/index.ts
@@ -1,5 +1,3 @@
 export * from './components/Clock';
-export * from './constants/ArduinoUno';
-export * from './constants/Constants';
 export * from './components/IOBindings';
 export * as Serial from './components/Serial';
