+++ src/simulator/assets/views/index.html
@@ -63,6 +63,5 @@
             Fork this on <a href="https://github.com/buzcarter/bounce-back-railway">GitHub</a>
         </footer>
     </main>
-    <script src="../../../../dist/trolley-controller.bundle.js"></script>
 </body>
 </html>
