+++ src/simulator/constants/index.ts
@@ -1,3 +1,2 @@
-export * from './Constants';
 export * from './CSSClasses';
 export * from './ElementIds';
